push init

+++ src/Controller/TaskController.php
@@ -6,23 +6,23 @@ use App\Entity\Task;
 use App\Form\TaskType;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Doctrine\ORM\EntityManagerInterface;
-
 
 class TaskController extends AbstractController
 {
     
     #[Route('/tasks', name: 'task_list')]
-    public function listAction()
+    public function listAction(EntityManagerInterface $em)
     {
-        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findAll()]);
+        $tasks = $em->getRepository(Task::class)->findAll();
+        
+        return $this->render('task/list.html.twig', [
+            'tasks' => $tasks
+        ]);
     }
 
-   
     #[Route('/tasks/create', name: 'task_create')]
-
     public function createAction(Request $request, EntityManagerInterface $em)
     {
         $task = new Task();
@@ -31,6 +31,9 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            // ✅ Assigner l'utilisateur connecté comme auteur
+            $task->setAuthor($this->getUser());
+            
             $em->persist($task);
             $em->flush();
 
@@ -39,20 +42,20 @@ class TaskController extends AbstractController
             return $this->redirectToRoute('task_list');
         }
 
-        return $this->render('task/create.html.twig', ['form' => $form->createView()]);
+        return $this->render('task/create.html.twig', [
+            'form' => $form->createView()
+        ]);
     }
 
-    
     #[Route('/tasks/{id}/edit', name: 'task_edit')]
-    
-    public function editAction(Task $task, Request $request)
+    public function editAction(Task $task, Request $request, EntityManagerInterface $em)
     {
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $em->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
 
@@ -65,25 +68,24 @@ class TaskController extends AbstractController
         ]);
     }
 
-    
     #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
-
-    public function toggleTaskAction(Task $task)
+    public function toggleTaskAction(Task $task, EntityManagerInterface $em)
     {
         $task->toggle(!$task->isDone());
-        $this->getDoctrine()->getManager()->flush();
+        $em->flush();
 
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
 
         return $this->redirectToRoute('task_list');
     }
 
-   
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
-
-    public function deleteTaskAction(Task $task)
+    public function deleteTaskAction(Task $task, EntityManagerInterface $em)
     {
-        $em = $this->getDoctrine()->getManager();
+        // Vérification de l'auteur de la tâche
+        if ($task->getAuthor()->getId() !== $this->getUser()->getId()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez pas supprimer cette tâche.');
+        }
         $em->remove($task);
         $em->flush();
